Rename and move Patient files
